rename: api 요청 코드 model -> api
