+++ src/entities/shared-game/api/use-shared-games.query.ts
